+++ classes/book.rb
@@ -4,7 +4,7 @@ class Book < Item
   attr_accessor :cover_state, :publisher
 
   def initialize(publisher, cover_state, publish_date)
-    super(publish_date)
+    super(id, publish_date)
     @publisher = publisher
     @cover_state = cover_state
   end
