+++ data/books.json
@@ -1,1 +1,1 @@
-[{"publisher":"jojo","publish_date":"2020/05/12","cover_state":"good"},{"publisher":"jdjd","publish_date":"2020/05/14","cover_state":"bad"}]
+[{"publisher":"jk","publish_date":"12/09/8","cover_state":"good"}]
