+++ data/labels.json
@@ -1,1 +1,1 @@
-[{"title":"fhfhf","color":"red"},{"title":"fghf","color":"orange"}]
+[{"title":"ju","color":"red"}]
